+++ package.json
@@ -1,17 +1,22 @@
 {
   "private": true,
   "name": "slacking-off",
-  "version": "0.1.2",
-  "description": "Slack support server. Slash commands!",
-  "main": "index.js",
+  "version": "1.0.0",
+  "description": "Bullshit slash commands for Slack!",
+  "main": "build/index.js",
   "scripts": {
-    "start": "node index.js"
+    "build": "rm -rf build && tsc",
+    "start": "node build/index.js"
   },
   "dependencies": {
+    "@types/express": "^4.17.1",
+    "@types/node": "^12.7.2",
+    "@types/request-promise": "^4.1.44",
     "body-parser": "^1.19.0",
     "express": "^4.16.4",
     "request": "^2.88.0",
-    "request-promise": "^4.2.4"
+    "request-promise": "^4.2.4",
+    "typescript": "^3.5.3"
   },
   "engines": {
     "node": "12.4.0"
