+++ src/commands/digital.ts
@@ -1,5 +1,7 @@
+import { RequestHandler } from 'express'
+
 // even out spacing, uppercase, and return in 'code' format
-const digitizer = digitalText => {
+const digitizer = (digitalText: string) => {
   digitalText = digitalText
     .split(' ')
     .join('')
@@ -9,13 +11,13 @@ const digitizer = digitalText => {
   return '`' + digitalText + '`'
 }
 
-module.exports = (req, res) => {
+export const digital: RequestHandler = ({ body }, { send }) => {
   const defaultText = '`D I G I T A L S P O R T S`'
 
   const result = {
     response_type: 'in_channel',
-    text: req.body.text ? digitizer(req.body.text) : defaultText
+    text: body.text ? digitizer(body.text) : defaultText
   }
 
-  res.send(result)
+  send(result)
 }
