+++ src/commands/pitchfork.ts
@@ -1,0 +1,18 @@
+import { RequestHandler } from 'express'
+
+const shitpost = `
+  ANGRY AT OP? WANT TO JOIN THE MOB? _WE HAVE GOT YOU COVERED!_
+  Traditional:  ---E
+  Left Handed:  Ǝ---
+  Fancy:        ---{
+`
+
+// Pitchform Emporioum
+export const pitchfork: RequestHandler = (_, { send }) => {
+  const result = {
+    response_type: 'in_channel',
+    text: shitpost
+  }
+
+  send(result)
+}
